backend/requirements.txt
# Web framework
fastapi==0.112.0
uvicorn[standard]==0.30.3

# Database
sqlalchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# Config
pydantic-settings==2.3.4

# Groq
groq==0.9.0

# PDF parsing
pypdf==4.3.1

# Multipart uploads
python-multipart==0.0.9

# Dev/test
pytest==8.3.2
pytest-asyncio==0.23.8
httpx==0.27.0
